+++ engine/range.h
@@ -15,6 +15,9 @@ extern "C" {
 void AspGetRange
     (AspEngine *engine, const AspDataEntry *entry,
      int32_t *startValue, int32_t *endValue, int32_t *stepValue);
+void AspGetSliceRange
+    (AspEngine *engine, const AspDataEntry *entry, int32_t sequenceCount,
+     int32_t *startValue, int32_t *endValue, int32_t *stepValue);
 bool AspIsValueAtRangeEnd
     (int32_t testValue, int32_t endValue, int32_t stepValue);