+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changes
 
 Version 0.7.1.8 (compiler 0.7.1.3, engine 0.7.3.7):
 - Engine:
+  - Fixed slicing issues with slice components of mixed sign.
   - Fixed a bug with assignment to list slices. Assignment to tuple slices was
     unaffected.